fix(transfer-detection): apply PR3 consolidated code-review — CODE_GUIDE §2 file/extension organization
