+++ Shared/TransferDetection/FuzzyTransferDetector.swift
@@ -1,13 +1,5 @@
 import Foundation
 
-/// One detected fuzzy candidate pair. Ordered: `newlyImported` is the
-/// just-imported transaction; `existingCounterpart` is the match it
-/// pairs with (which may itself be in the newly-imported set).
-struct TransferCandidatePair: Sendable, Hashable {
-  let newlyImported: Transaction
-  let existingCounterpart: Transaction
-}
-
 /// Pure fuzzy candidate finder. No I/O.
 ///
 /// Eligibility is determined by `Transaction.transferDetectionValueLeg`.
